Scott
The Python script must run in the same directory as the Label files. It is okay to have other files within that are not related to labels as long as they are not .txt. To run the code type main() into the shell. The result will be a 'Scott.csv' which contains instances where there is space/overlap in between TA-Q: and the next TA events.